patching IR (tested on HW)
