+++ libs/pulse/pxt.json
@@ -6,7 +6,6 @@
         "pulse.cpp",
         "pulse.h",
         "bitvector.h",
-        "samd21.cpp",
         "shims.d.ts",
         "ns.ts"
     ],
